+++ a2a/core.go
@@ -35,6 +35,14 @@ type TaskInfoProvider interface {
 	TaskInfo() TaskInfo
 }
 
+// MetadataCarrier provides access to extensions metadata container.
+type MetadataCarrier interface {
+	// Meta returns the metadata container.
+	Meta() map[string]any
+	// SetMeta sets the metadata value for the provided key.
+	SetMeta(k string, v any)
+}
+
 // TaskInfo represents information about the Task and the group of interactions it belongs to.
 // Values might be empty which means the TaskInfoProvider is not associated with any tasks.
 // An example would be the first user message.
@@ -64,11 +72,9 @@ func (*Message) isSendMessageResult() {}
 // Event interface is used to represent types that can be sent over a streaming connection.
 type Event interface {
 	TaskInfoProvider
+	MetadataCarrier
 
 	isEvent()
-
-	// Meta returns root Metadata of the struct which implements Event.
-	Meta() map[string]any
 }
 
 func (*Message) isEvent()                 {}
@@ -199,6 +205,10 @@ func (m *Message) Meta() map[string]any {
 	return m.Metadata
 }
 
+func (m *Message) SetMeta(k string, v any) {
+	setMeta(&m.Metadata, k, v)
+}
+
 func (m *Message) TaskInfo() TaskInfo {
 	return TaskInfo{TaskID: m.TaskID, ContextID: m.ContextID}
 }
@@ -320,6 +330,10 @@ func (m *Task) Meta() map[string]any {
 	return m.Metadata
 }
 
+func (m *Task) SetMeta(k string, v any) {
+	setMeta(&m.Metadata, k, v)
+}
+
 func (m *Task) TaskInfo() TaskInfo {
 	return TaskInfo{TaskID: m.ID, ContextID: m.ContextID}
 }
@@ -351,6 +365,14 @@ type Artifact struct {
 
 	// Parts is an array of content parts that make up the artifact.
 	Parts ContentParts `json:"parts" yaml:"parts" mapstructure:"parts"`
+}
+
+func (a *Artifact) Meta() map[string]any {
+	return a.Metadata
+}
+
+func (a *Artifact) SetMeta(k string, v any) {
+	setMeta(&a.Metadata, k, v)
 }
 
 var _ Event = (*TaskArtifactUpdateEvent)(nil)
@@ -389,6 +411,10 @@ func (e TaskArtifactUpdateEvent) MarshalJSON() ([]byte, error) {
 
 func (a *TaskArtifactUpdateEvent) Meta() map[string]any {
 	return a.Metadata
+}
+
+func (a *TaskArtifactUpdateEvent) SetMeta(k string, v any) {
+	setMeta(&a.Metadata, k, v)
 }
 
 func (m *TaskArtifactUpdateEvent) TaskInfo() TaskInfo {
@@ -469,6 +495,10 @@ func NewStatusUpdateEvent(infoProvider TaskInfoProvider, state TaskState, msg *M
 
 func (a *TaskStatusUpdateEvent) Meta() map[string]any {
 	return a.Metadata
+}
+
+func (a *TaskStatusUpdateEvent) SetMeta(k string, v any) {
+	setMeta(&a.Metadata, k, v)
 }
 
 func (m *TaskStatusUpdateEvent) TaskInfo() TaskInfo {
@@ -560,6 +590,10 @@ func (p TextPart) Meta() map[string]any {
 	return p.Metadata
 }
 
+func (p *TextPart) SetMeta(k string, v any) {
+	setMeta(&p.Metadata, k, v)
+}
+
 func (p TextPart) MarshalJSON() ([]byte, error) {
 	type wrapped TextPart
 	type withKind struct {
@@ -580,6 +614,10 @@ type DataPart struct {
 
 func (p DataPart) Meta() map[string]any {
 	return p.Metadata
+}
+
+func (p *DataPart) SetMeta(k string, v any) {
+	setMeta(&p.Metadata, k, v)
 }
 
 func (p DataPart) MarshalJSON() ([]byte, error) {
@@ -603,6 +641,10 @@ type FilePart struct {
 
 func (p FilePart) Meta() map[string]any {
 	return p.Metadata
+}
+
+func (p *FilePart) SetMeta(k string, v any) {
+	setMeta(&p.Metadata, k, v)
 }
 
 func (p FilePart) MarshalJSON() ([]byte, error) {
@@ -692,6 +734,14 @@ type TaskIDParams struct {
 	Metadata map[string]any `json:"metadata,omitempty" yaml:"metadata,omitempty" mapstructure:"metadata,omitempty"`
 }
 
+func (p *TaskIDParams) Meta() map[string]any {
+	return p.Metadata
+}
+
+func (p *TaskIDParams) SetMeta(k string, v any) {
+	setMeta(&p.Metadata, k, v)
+}
+
 // TaskQueryParams defines parameters for querying a task, with an option to limit history length.
 type TaskQueryParams struct {
 	// HistoryLength is the number of most recent messages from the task's history to retrieve.
@@ -702,6 +752,14 @@ type TaskQueryParams struct {
 
 	// Metadata is an optional metadata associated with the request.
 	Metadata map[string]any `json:"metadata,omitempty" yaml:"metadata,omitempty" mapstructure:"metadata,omitempty"`
+}
+
+func (p *TaskQueryParams) Meta() map[string]any {
+	return p.Metadata
+}
+
+func (p *TaskQueryParams) SetMeta(k string, v any) {
+	setMeta(&p.Metadata, k, v)
 }
 
 // MessageSendConfig defines configuration options for a `message/send` or `message/stream` request.
@@ -731,6 +789,21 @@ type MessageSendParams struct {
 
 	// Metadata is an optional metadata for extensions.
 	Metadata map[string]any `json:"metadata,omitempty" yaml:"metadata,omitempty" mapstructure:"metadata,omitempty"`
+}
+
+func (p *MessageSendParams) Meta() map[string]any {
+	return p.Metadata
+}
+
+func (p *MessageSendParams) SetMeta(k string, v any) {
+	setMeta(&p.Metadata, k, v)
+}
+
+func setMeta(m *map[string]any, k string, v any) {
+	if *m == nil {
+		*m = make(map[string]any)
+	}
+	(*m)[k] = v
 }
 
 // Time-based UUID generally improves index update performance if ID field is indexed in a persistent store.