+++ a2asrv/handler.go
@@ -69,6 +69,7 @@ type RequestHandler interface {
 type defaultRequestHandler struct {
 	agentExecutor AgentExecutor
 	execManager   taskexec.Manager
+	panicHandler  taskexec.PanicHandlerFn
 
 	pushSender        PushSender
 	queueManager      eventqueue.Manager
@@ -101,6 +102,13 @@ func WithLogger(logger *slog.Logger) RequestHandlerOption {
 func WithEventQueueManager(manager eventqueue.Manager) RequestHandlerOption {
 	return func(ih *InterceptedHandler, h *defaultRequestHandler) {
 		h.queueManager = manager
+	}
+}
+
+// WithExecutionPanicHandler allows to set a custom handler for panics occurred during execution.
+func WithExecutionPanicHandler(handler func(r any) error) RequestHandlerOption {
+	return func(ih *InterceptedHandler, h *defaultRequestHandler) {
+		h.panicHandler = handler
 	}
 }
 
@@ -153,6 +161,7 @@ func NewHandler(executor AgentExecutor, options ...RequestHandlerOption) Request
 			QueueManager:      h.queueManager,
 			ConcurrencyConfig: h.concurrencyConfig,
 			Factory:           execFactory,
+			PanicHandler:      h.panicHandler,
 		})
 	} else {
 		if h.queueManager == nil {
@@ -166,6 +175,7 @@ func NewHandler(executor AgentExecutor, options ...RequestHandlerOption) Request
 			QueueManager:      h.queueManager,
 			ConcurrencyConfig: h.concurrencyConfig,
 			Factory:           execFactory,
+			PanicHandler:      h.panicHandler,
 		})
 	}